refactor!: restructure whole project for new microservice integration
