+++ notification-service/.golangci.yml
