+++ notification-service/go.mod
@@ -1,0 +1,5 @@
+module notification-service
+
+go 1.22
+
+require github.com/rabbitmq/amqp091-go v1.10.0 // indirect
