+++ rate-service/docs/docs.go
