+++ rate-service/docs/hw5-application-architecture.png
