+++ rate-service/docs/swagger.json
