+++ rate-service/docs/swagger.yaml
