+++ rate-service/go.mod
@@ -1,4 +1,4 @@
-module github.com/GenesisEducationKyiv/software-engineering-school-4-0-DimaL-cloud
+module rate-service
 
 go 1.22
