+++ rate-service/internal/client/client.go
