+++ rate-service/internal/client/rate/fawazahmed.go
