+++ rate-service/internal/client/rate/privatbank.go
