+++ rate-service/internal/client/rate/rate.go
