+++ rate-service/internal/configs/mail.go
