+++ rate-service/internal/handler/handler.go
@@ -1,10 +1,10 @@
 package handler
 
 import (
-	"github.com/GenesisEducationKyiv/software-engineering-school-4-0-DimaL-cloud/internal/service"
 	"github.com/gin-gonic/gin"
 	swaggerFiles "github.com/swaggo/files"
 	ginSwagger "github.com/swaggo/gin-swagger"
+	"rate-service/internal/service"
 )
 
 type Handler struct {
