+++ rate-service/internal/handler/rate.go
