+++ rate-service/internal/handler/response.go
