+++ rate-service/internal/models/errors.go
