+++ rate-service/internal/models/subscription.go
