+++ rate-service/internal/repository/subscription.go
@@ -3,9 +3,9 @@ package repository
 import (
 	"errors"
 	"fmt"
-	"github.com/GenesisEducationKyiv/software-engineering-school-4-0-DimaL-cloud/internal/models"
 	"github.com/jmoiron/sqlx"
 	"github.com/lib/pq"
+	"rate-service/internal/models"
 	"strings"
 )
