+++ rate-service/internal/service/mail.go
@@ -1,8 +1,8 @@
 package service
 
 import (
-	"github.com/GenesisEducationKyiv/software-engineering-school-4-0-DimaL-cloud/internal/configs"
 	"net/smtp"
+	"rate-service/internal/configs"
 )
 
 type Mail interface {
